+++ server/src/main/java/artgarden/server/entity/dto/reviewDto/ReviewDto.java
@@ -11,7 +11,7 @@ public class ReviewDto {
 
     private String perform_id;
     private String content;
-    private Long rate;
+    private Double rate;
     private Long member_id;
     private LocalDateTime created_at;
     private LocalDateTime modified_at;
